modify hero section
