+++ app/(pages)/events/page.tsx
@@ -2,6 +2,7 @@ import {  eventsData, footerProps } from '@/app/constants'
 import BackToTopButton from '@/components/BackToTop'
 import FooterSection from '@/components/Footer'
 import { HeroSection } from '@/components/Hero'
+import { HeroContent } from '@/components/HeroContent'
 import JoinUsSection from '@/components/JoinUs'
 import Navbar from '@/components/NavBar'
 import { UpcomingEvents } from '@/components/UpcomingEvents'
@@ -11,7 +12,9 @@ const page = () => {
   return (
     <div>
         <Navbar />
-        <HeroSection {...eventsData} />
+        <HeroSection>
+          <HeroContent {...eventsData} />
+        </HeroSection>
         <UpcomingEvents />
         <JoinUsSection />
         <FooterSection {...footerProps} />
